Create separate 'analyze' and 'run' commands
Change-type: minor
