+++ package.json
@@ -48,14 +48,15 @@
   "oclif": {
     "bin": "migrator",
     "dirname": "migrator",
-    "default": ".",
-    "commands": "./dist",
-    "plugins": [],
+    "commands": "./dist/commands",
+    "plugins": [
+      "@oclif/plugin-help"
+    ],
     "topicSeparator": " ",
     "topics": {}
   },
   "pkg": {
-    "scripts": "dist/*.js",
+    "scripts": "dist/**/*.js",
     "assets": [
       "node_modules/drivelist/build/Release/drivelist.node",
       "node_modules/lzma-native/prebuilds/**/*",
